rename 'modTypes.ts' to 'modDefinitionTypes.ts'
