+++ src/extensionConsole/pages/activateRecipe/ServicesBody.tsx
@@ -19,7 +19,7 @@ import styles from "./ServicesBody.module.scss";
 
 import React, { useMemo } from "react";
 import { Card } from "react-bootstrap";
-import { type ModDefinition } from "@/types/modTypes";
+import { type ModDefinition } from "@/types/modDefinitionTypes";
 import AuthWidget from "@/components/auth/AuthWidget";
 import ServiceDescriptor from "@/extensionConsole/pages/activateRecipe/ServiceDescriptor";
 import { useField } from "formik";
