+++ src/installables/installableTypes.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type ModDefinition } from "@/types/modTypes";
+import { type ModDefinition } from "@/types/modDefinitionTypes";
 import { type Organization } from "@/types/contract";
 import { type ResolvedExtension } from "@/types/extensionTypes";
 import { type RegistryId } from "@/types/registryTypes";
