+++ src/pageEditor/panes/save/useSavingWizard.test.tsx
@@ -38,7 +38,7 @@ import pDefer from "p-defer";
 import { pick } from "lodash";
 import extensionsSlice from "@/store/extensionsSlice";
 import { getMinimalUiSchema } from "@/components/formBuilder/formBuilderHelpers";
-import { type OptionsDefinition } from "@/types/modTypes";
+import { type OptionsDefinition } from "@/types/modDefinitionTypes";
 import { useAllRecipes } from "@/recipes/recipesHooks";
 import { installedRecipeMetadataFactory } from "@/testUtils/factories/extensionFactories";
 import {
