+++ src/pageEditor/sidebar/ExtensionEntry.tsx
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { type ModDefinition } from "@/types/modTypes";
+import { type ModDefinition } from "@/types/modDefinitionTypes";
 import React from "react";
 import { type FormState } from "@/pageEditor/extensionPoints/formStateTypes";
 import { isExtension } from "./common";
