+++ src/pageEditor/slices/editorSlice.ts
@@ -91,7 +91,7 @@ import {
 } from "@/components/fields/schemaFields/serviceFieldUtils";
 import { type UUID } from "@/types/stringTypes";
 import { type RegistryId } from "@/types/registryTypes";
-import { type OptionsDefinition } from "@/types/modTypes";
+import { type OptionsDefinition } from "@/types/modDefinitionTypes";
 import { type IExtension } from "@/types/extensionTypes";
 import { type OptionsArgs } from "@/types/runtimeTypes";
