+++ src/store/extensionsSlice.ts
@@ -38,7 +38,10 @@ import {
   selectSourceRecipeMetadata,
 } from "@/types/extensionTypes";
 import { type UUID } from "@/types/stringTypes";
-import { type ModDefinition, type ExtensionDefinition } from "@/types/modTypes";
+import {
+  type ModDefinition,
+  type ExtensionDefinition,
+} from "@/types/modDefinitionTypes";
 import { type RegistryId } from "@/types/registryTypes";
 import { type OutputKey, type OptionsArgs } from "@/types/runtimeTypes";
