+++ src/types/modDefinitionTypes.ts
